Final version of code with bug fixes
The final version of the code. Has been tested on my personal machine and is verified to work. I forgot to commit some changes before going bug hunting (oops) and lost track of what was changed so instead I simply used the good ol' ctrl+C  and ctrl+V. nothing structurally is different just had some issues with the pathing that was fixed.